move route function to `route.py`
